Fix compliation issue on 32 bit systems
Cast to unsigned long before casting to pointer.
Problem introduced in fa57e32.

+++ drivers/infiniband/hw/mlx5/mr.c
@@ -644,7 +644,7 @@ void mlx5_umr_cq_handler(struct ib_cq *cq, void *cq_context)
 		if (err == 0)
 			break;
 
-		mr = (struct mlx5_ib_mr *)wc.wr_id;
+		mr = (struct mlx5_ib_mr *)(unsigned long)wc.wr_id;
 		mr->status = wc.status;
 		complete(&mr->done);
 	}
